Multiplayer and menu update
New multiplayer protocol and start or rewriting menu to final state machine

+++ Client/GameLogic.h
@@ -11,7 +11,7 @@
 
 struct SnakeBlock
 {
-	int PhysicalBodyID;
+	short PhysicalBodyID;
 	char Direction;
 };
